Fix `S3Storage.delete()` when using preloading
When `AWS_PRELOAD_METADATA` is `True` the function
`S3Storage.exists()` looks in `S3Storage.entries` whether a file
exists. This dictionary of files is, however, not updated when calling
`S3Storage.delete()` and so `S3Storage.exists()` will return bogus
results if you use it on a file that was recently deleted.